Add docs to tree_printer functions, remove the Formatter wrapper for easier tree printing
