+++ tests/test_tree_printer.py
@@ -1,7 +1,6 @@
 from hyperiax.tree.builders import THeight_legacy
-from hyperiax.tree.utils import HypTreeFormatter
+from hyperiax.tree.printer_utils import HypTreeFormatter
 
 def test_tree_formatter():
-    tree = THeight_legacy(h=5, degree=3)
-    tree_formatter = HypTreeFormatter(tree)
-    tree_formatter.print_tree()
+    tree = THeight_legacy(h=4, degree=3)
+    tree.plot_tree_text()
